requirements.txt
numpy>=1.24.0
opencv-python>=4.8.0
matplotlib>=3.7.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.5.0
ipywidgets>=8.0.0
tqdm>=4.65.0
